--- Editor/Editors/CinemachineSplineRollEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 90a9d77aec8a7d2408cbec8d8d2f7be4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
